requirements.txt
watchdog==4.0.2
pyyaml==6.0.2
